requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
google-cloud-texttospeech==2.14.1
pyyaml==6.0